Updating onlinecop branch to match up with origin/master.

+++ include/enemyc.h
@@ -26,15 +26,9 @@
 
 int select_encounter(int, int);         /*  combat.c  */
 void enemy_init(void);          /*  combat.c  */
-void enemy_chooseaction(int);   /*  combat.c  */
-void enemy_charmaction(int);    /*  combat.c, heroc.c  */
+void enemy_chooseaction(size_t);   /*  combat.c  */
+void enemy_charmaction(size_t);    /*  combat.c, heroc.c  */
 
 
 #endif  /* __ENEMYC_H */
 
-/* Local Variables:     */
-/* mode: c              */
-/* comment-column: 0    */
-/* indent-tabs-mode nil */
-/* tab-width: 4         */
-/* End:                 */
