+++ include/movement.h
@@ -24,14 +24,8 @@
 #define __MOVEMENT_H 1
 
 
-int find_path(int, int, int, int, int, char *, int);
+int find_path(size_t, unsigned int, unsigned int, unsigned int, unsigned int, char *, unsigned int);
 
 
 #endif  /* __MOVEMENT_H */
 
-/* Local Variables:     */
-/* mode: c              */
-/* comment-column: 0    */
-/* indent-tabs-mode nil */
-/* tab-width: 4         */
-/* End:                 */
